+++ supporting-files/quarterly_processing/04-create-docker-images.sh
@@ -21,24 +21,24 @@ echo "--------------------------------------------------------------------------
 podman push docker.io/minus34/gnafloader:latest
 podman push docker.io/minus34/gnafloader:202302
 
-echo "---------------------------------------------------------------------------------------------------------------------"
-echo "build gnaf-loader GDA2020 docker image"
-echo "---------------------------------------------------------------------------------------------------------------------"
-
-podman build --tag docker.io/minus34/gnafloader:latest-gda2020 --tag docker.io/minus34/gnafloader:202302-gda2020 --build-arg BASE_URL="https://minus34.com/opendata/geoscape-202302-gda2020" .
-
-echo "---------------------------------------------------------------------------------------------------------------------"
-echo "push image (with 2 tags) to Docker Hub"
-echo "---------------------------------------------------------------------------------------------------------------------"
-
-# login (if needed)
-#podman login -u ${DOCKER_USER} -p ${DOCKER_PASSWORD} docker.io
-
-podman push docker.io/minus34/gnafloader:latest-gda2020
-podman push docker.io/minus34/gnafloader:202302-gda2020
+#echo "---------------------------------------------------------------------------------------------------------------------"
+#echo "build gnaf-loader GDA2020 docker image"
+#echo "---------------------------------------------------------------------------------------------------------------------"
+#
+#podman build --tag docker.io/minus34/gnafloader:latest-gda2020 --tag docker.io/minus34/gnafloader:202302-gda2020 --build-arg BASE_URL="https://minus34.com/opendata/geoscape-202302-gda2020" .
+#
+#echo "---------------------------------------------------------------------------------------------------------------------"
+#echo "push image (with 2 tags) to Docker Hub"
+#echo "---------------------------------------------------------------------------------------------------------------------"
+#
+## login (if needed)
+##podman login -u ${DOCKER_USER} -p ${DOCKER_PASSWORD} docker.io
+#
+#podman push docker.io/minus34/gnafloader:latest-gda2020
+#podman push docker.io/minus34/gnafloader:202302-gda2020
 
 #echo "---------------------------------------------------------------------------------------------------------------------"
 #echo "clean up Docker locally - warning: this could accidentally destroy other Docker images"
 #echo "---------------------------------------------------------------------------------------------------------------------"
 #
-#echo 'y' | docker system prune
+#echo 'y' | podman system prune